# feat: faster wrapping