--- a/.eslintrc
+++ b/.eslintrc
@@ -20,7 +20,7 @@
 
   rules: {
     // Best Practices
-    accessor-pairs: error,
+    accessor-pairs: off,
     array-callback-return: error,
     block-scoped-var: error,
     class-methods-use-this: off,
@@ -66,9 +66,10 @@
     no-octal-escape: error,
     no-param-reassign: off,
     no-proto: error,
-    no-redeclare: error,
+    no-redeclare: off,
+    "@typescript-eslint/no-redeclare": ["error"],
     no-restricted-properties: error,
-    no-return-assign: error,
+    no-return-assign: off,
     no-return-await: error,
     no-script-url: error,
     no-self-assign: error,