--- a/test/MappingIterator-test.js
+++ b/test/MappingIterator-test.js
@@ -542,7 +542,6 @@ describe('MappingIterator', () => {
 
   describe('A chain of maps and filters', () => {
     for (const iteratorGen of [() => range(0, 2), () => fromArray([0, 1, 2]), () => wrap(range(0, 2))]) {
-      // eslint-disable-next-line no-loop-func
       describe(`with ${iteratorGen()}`, () => {
         let iterator;