This PR covers step 2. in the plan described in https://github.com/RubenVerborgh/AsyncIterator/issues/44#issuecomment-1086950726 and aims to 1) lower the overhead of the `wrap()` function and 2) expand support to additional types of sources (namely `Iterator` and `Iterable` objects).

The rationale for this PR is taking the onus of figuring out how to best wrap common types of data sources away from users of this library. Users should be able to expect `wrap()` to use the strategy that is most appropriate for any given source (as long as it's a supported source).

Performance-wise, this PR does two things:

1) It introduces source-specific iterators that synchronously read from their sources, skipping the internal buffering done by `TransformIterator` (i.e. the class currently used for wrapping)

2) It introduces a specific option that allows `AsyncIterator` instances to fall-through as they are, with no wrapping applied

The following test, admittedly naive, takes ~116 ms with the current `main` branch and ~8 ms with this PR (`letIteratorThrough` makes virtually no difference). This mostly highlight the difference made by skipping `TransformIterator`'s internal buffering.

```typescript
let i = 0;
const arr = new Array(200_000).fill(true).map(() => i++);
const source = new ArrayIterator(arr);
const opts = { letIteratorThrough: false };
const then = Date.now();
wrap(source, opts).on('data', () => {}).on('end', () => {
  console.log('elapsed', Date.now() - then, 'ms');
});
```

Things become much more interesting when we consider what happens with a chain established by `wrap()`ing:

```typescript
let i = 0;
const arr = new Array(200_000).fill(true).map(() => i++);
const source = new ArrayIterator(arr);
const opts = { letIteratorThrough: false };
const then = Date.now();
wrap(wrap(wrap(wrap(source, opts), opts), opts), opts).on('data', () => {}).on('end', () => {
  console.log('elapsed', Date.now() - then, 'ms');
});
```

The above test takes ~250 ms with the current `main` branch, ~ 15 ms with this PR and `letIteratorThrough: false`, ~ 8 ms with this PR and `letIteratorThrough: true`. I haven't seen consistent differences with or without warmups.

My use case is the optimization of the interface between [quadstore](https://github.com/belayeng/quadstore), which implements the RDF/JS `Source` interface, and [comunica](https://comunica.dev), through which the former can be queried via SPARQL: https://github.com/comunica/comunica/blob/a2f31270778730911bd0bdf8b8c7ade567153319/packages/actor-rdf-resolve-quad-pattern-rdfjs-source/lib/RdfJsQuadSource.ts#L29 .

As many algebraic operation (matching, joining, ...) are happening in-memory, thus requiring the streaming of large amounts of quads from the persistence layer into the engine, removing even just one layer of indirection/buffering can make a significant difference.