@jacoscaz 2022-04-07
I should add that this PR is not ready to be merged as it's lacking tests and I'm sure the code can be improved. However, it is ready for first round of feedbacks to better direct my efforts. Tagging @jeswr , @rubensworks and @RubenVerborgh to keep everyone in the loop. BTW, this would have to be rebased on `main` once #59 gets merged.