@jacoscaz 2022-04-08
@jeswr I've tweaked your test a bit to ignore the time spent establishing the chain, just in case it could make a difference (likely insignificant, but still).

```typescript
let i = 0;
const arr = new Array(200_000).fill(true).map(() => i++);

const source = new ArrayIterator(arr);
const opts = { /* letIteratorThrough: true */ };

const chain = wrap(Promise.resolve(wrap(Promise.resolve(wrap(Promise.resolve(wrap(Promise.resolve(source), opts)), opts)), opts)), opts);

chain.once('readable', () => {
  const then = Date.now();
  chain.on('data', () => {}).on('end', () => {
    console.log('elapsed', Date.now() - then, 'ms');
  });
});
```

This test runs in ~450 ms with current `main` and ~12 ms with this PR. `letIteratorThrough` makes no difference as having to return an iterator _synchronously_ forces the use of `WrappingIterator` even for `AsyncIterator` sources as we still need to wait for the promise to asynchronously resolve.