@jacoscaz 2022-04-14
Not sure about what's going on with coveralls but this is probably ready for review by @RubenVerborgh (in due time, after #59 is merged). I'll start working on cloning.