@jacoscaz 2022-07-04
@RubenVerborgh I have just rebased this one on `main`. I'm not sure as to what is happening with the failing test on Node 10.x but I would prefer if you were to take a look first before doing further work on this.