@jeswr 2022-07-28
> Thanks, removed the second option https://github.com/RubenVerborgh/AsyncIterator/commit/16fcfb638a97df3b26cb11aa3d4da060b120c146 and will talk to Jesse over lunch about the other one!

The request for that was in relation to performance when wrapping N3.js.

By default `wrap(store.match(...))` would end up using `fromAsyncReadable` which in turn uses buffering within the `WrappingIterator` (and all results get buffered into an array as a result of [this](https://github.com/rdfjs/N3.js/blob/b2661908d8af3e57fcb67111b32752118a8590d0/src/N3Store.js#L795-L799)). As opposed to just using the `IterableWrapper` which is much faster as it just pulls out quads on-demand. IIRC we cannot change the default prioritization order for backwards compatibility reasons.

The main consequence of removing this is that downstream clients like Comunica will sometimes need to write 

```ts
 if ((isIterable(source) || isIterator(source)))
    return fromIterable<T>(source);
 else if
    return wrap<T>(source);
```

rather than
```ts
return wrap(source, { prioritizeIterable: true })
```