@jacoscaz 2022-07-29
> Thanks @jeswr; I'll take a look at the backwards compat thing too then. But we should def put performance-friendly options first. (Also this will be one of the last changes before semver.major, so then we can happily break.)

Thanks @jeswr for the refresher. I hadn't realized we already have a use case for the `prioritizeIterable` flag. Long-term, I suspect more such flags might follow (e.g. for [native async iteration](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Iteration_protocols#the_async_iterator_and_async_iterable_protocols)). Perhaps an explicit ordering might be better, here? Something like `priority: ['iterable', 'asynciterable', 'array']`?