@RubenVerborgh 2022-07-31
(Just fixing a small mistake in 88165c48d15972fb0dc27bd81d2149562014264f that causes CI failures.)