@RubenVerborgh 2022-07-31
Published as v3.6.0 🥳